requirements.txt
keras==2.15.0
numpy==1.26.0
opencv_python==4.8.1.78
Pillow==10.1.0
Requests==2.31.0
streamlit==1.28.2
tensorflow==2.15.0
tensorflow_intel==2.15.0
